# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Controle da Operação API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginRequest {
  /** Email or username */
  login: string;
  password: string;
}

export type UserInfoRole = (typeof UserInfoRole)[keyof typeof UserInfoRole];

export const UserInfoRole = {
  master: "master",
  admin: "admin",
  user: "user",
} as const;

export interface UserPermissions {
  fluxoDados: boolean;
  fluxoCancelamento: boolean;
  fluxoEmissao: boolean;
}

export interface UserInfo {
  id: number;
  name: string;
  email: string;
  username: string;
  role: UserInfoRole;
  permissions: UserPermissions;
}

export interface AuthResponse {
  user: UserInfo;
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  master: "master",
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  username: string;
  role: UserRole;
  active: boolean;
  permissions: UserPermissions;
  createdAt: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  user: "user",
} as const;

export interface CreateUserRequest {
  name: string;
  email: string;
  username: string;
  password: string;
  role: CreateUserRequestRole;
  permissions?: UserPermissions;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  user: "user",
} as const;

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  username?: string;
  password?: string;
  role?: UpdateUserRequestRole;
  active?: boolean;
}

export interface UserPermissionsRequest {
  fluxoDados: boolean;
  fluxoCancelamento: boolean;
  fluxoEmissao: boolean;
}

export interface Order {
  id: number;
  orderNumber: string;
  supplier: string;
  product: string;
  startDate?: string;
  booking?: string;
  emissionCost?: number;
  paxTotal?: number;
  agency?: string;
  issuedBy?: string;
  status: string;
  deliveryStatus: string;
  notes?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface CreateOrderRequest {
  orderNumber: string;
  supplier: string;
  product: string;
  startDate?: string;
  booking?: string;
  emissionCost?: number;
  paxTotal?: number;
  agency?: string;
  issuedBy?: string;
  notes?: string;
}

export interface UpdateOrderRequest {
  supplier?: string;
  product?: string;
  startDate?: string;
  booking?: string;
  emissionCost?: number;
  paxTotal?: number;
  agency?: string;
  issuedBy?: string;
  status?: string;
  deliveryStatus?: string;
  notes?: string;
}

export interface OrderListResponse {
  data: Order[];
  total: number;
  page: number;
  pageSize: number;
}

export interface ImportCsvRequest {
  csvContent: string;
  separator?: string;
}

export interface ImportCsvResponse {
  imported: number;
  skipped: number;
  errors: string[];
  message: string;
}

export interface Cancellation {
  id: number;
  orderNumber: string;
  supplier?: string;
  product?: string;
  booking?: string;
  value?: number;
  useDate?: string;
  pax?: number;
  reason: string;
  status: string;
  createdBy: string;
  sendDate?: string;
  solutionDate?: string;
  daysPending?: number;
  emailSent: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface CreateCancellationRequest {
  orderNumber: string;
  supplier?: string;
  product?: string;
  booking?: string;
  value?: number;
  useDate?: string;
  pax?: number;
  reason: string;
  notes?: string;
}

export interface UpdateCancellationRequest {
  status?: string;
  reason?: string;
  pax?: number;
  solutionDate?: string;
  notes?: string;
}

export interface CancellationStats {
  pending: number;
  resolved: number;
  emissionError: number;
  checkError: number;
}

export interface CancellationListResponse {
  data: Cancellation[];
  total: number;
  page: number;
  pageSize: number;
  stats?: CancellationStats;
}

export interface Goal {
  id: number;
  userId: number;
  userName: string;
  userEmail: string;
  date: string;
  target: number;
  achieved?: number;
  percentage?: number;
  createdAt: string;
}

export interface CreateGoalRequest {
  userId: number;
  date: string;
  target: number;
  achieved?: number;
}

export interface PerformanceEntry {
  date: string;
  userName: string;
  target: number;
  achieved: number;
  percentage: number;
}

export interface PerformanceResponse {
  data: PerformanceEntry[];
  totalTarget: number;
  totalAchieved: number;
  overallPercentage: number;
}

export interface TopPerformer {
  userName: string;
  achieved: number;
  percentage: number;
}

export interface GoalsSummary {
  todayTarget: number;
  todayAchieved: number;
  monthTarget: number;
  monthAchieved: number;
  topPerformers: TopPerformer[];
}

export type ListOrdersParams = {
  search?: string;
  status?: string;
  page?: number;
  pageSize?: number;
};

export type ListCancellationsParams = {
  search?: string;
  status?: string;
  page?: number;
  pageSize?: number;
};

export type ListGoalsParams = {
  userId?: number;
  dateFrom?: string;
  dateTo?: string;
};

export type GetPerformanceParams = {
  userId?: number;
  dateFrom?: string;
  dateTo?: string;
};