# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Controle da Operação API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  /** Email or username */
  login: string;
  password: string;
}